refactor: update deps

+++ object_test.go
@@ -12,6 +12,8 @@ import (
 )
 
 func TestObjectMethodCall(t *testing.T) {
+	t.Parallel()
+
 	ctx := v8.NewContext()
 	iso := ctx.Isolate()
 	defer iso.Dispose()
@@ -44,9 +46,13 @@ func TestObjectMethodCall(t *testing.T) {
 }
 
 func TestObjectSet(t *testing.T) {
+	t.Parallel()
+
 	ctx := v8.NewContext()
 	defer ctx.Isolate().Dispose()
 	defer ctx.Close()
+
+	symIter := v8.SymbolIterator(ctx.Isolate())
 	val, _ := ctx.RunScript("const foo = {}; foo", "")
 	obj, _ := val.AsObject()
 	if err := obj.Set("bar", "baz"); err != nil {
@@ -71,6 +77,9 @@ func TestObjectSet(t *testing.T) {
 	}
 	if err := obj.Set("a", 0); err == nil {
 		t.Error("expected error but got <nil>")
+	}
+	if err := obj.SetSymbol(symIter, "sym"); err != nil {
+		t.Errorf("unexpected error: %v", err)
 	}
 	if err := obj.SetIdx(10, "ten"); err != nil {
 		t.Errorf("unexpected error: %v", err)
@@ -131,16 +140,25 @@ func TestObjectInternalFields(t *testing.T) {
 }
 
 func TestObjectGet(t *testing.T) {
+	t.Parallel()
+
 	ctx := v8.NewContext()
 	defer ctx.Isolate().Dispose()
 	defer ctx.Close()
-	val, _ := ctx.RunScript("const foo = { bar: 'baz'}; foo", "")
+	symIter := v8.SymbolIterator(ctx.Isolate())
+	val, err := ctx.RunScript("const foo = { bar: 'baz', [Symbol.iterator]: 'gee'}; foo", "")
+	if err != nil {
+		t.Fatalf("RunScript failed: %v", err)
+	}
 	obj, _ := val.AsObject()
 	if bar, _ := obj.Get("bar"); bar.String() != "baz" {
 		t.Errorf("unexpected value: %q", bar)
 	}
 	if baz, _ := obj.Get("baz"); !baz.IsUndefined() {
 		t.Errorf("unexpected value: %q", baz)
+	}
+	if got, _ := obj.GetSymbol(symIter); got.String() != "gee" {
+		t.Errorf("unexpected value: %q", got)
 	}
 	ctx.RunScript("foo[5] = 5", "")
 	if five, _ := obj.GetIdx(5); five.Integer() != 5 {
@@ -152,16 +170,22 @@ func TestObjectGet(t *testing.T) {
 }
 
 func TestObjectHas(t *testing.T) {
+	t.Parallel()
+
 	ctx := v8.NewContext()
 	defer ctx.Isolate().Dispose()
 	defer ctx.Close()
-	val, _ := ctx.RunScript("const foo = {a: 1, '2': 2}; foo", "")
+	symIter := v8.SymbolIterator(ctx.Isolate())
+	val, _ := ctx.RunScript("const foo = {a: 1, '2': 2, [Symbol.iterator]: 3}; foo", "")
 	obj, _ := val.AsObject()
 	if !obj.Has("a") {
 		t.Error("expected true, got false")
 	}
 	if obj.Has("c") {
 		t.Error("expected false, got true")
+	}
+	if !obj.HasSymbol(symIter) {
+		t.Error("expected true, got false")
 	}
 	if !obj.HasIdx(2) {
 		t.Error("expected true, got false")
@@ -172,10 +196,13 @@ func TestObjectHas(t *testing.T) {
 }
 
 func TestObjectDelete(t *testing.T) {
+	t.Parallel()
+
 	ctx := v8.NewContext()
 	defer ctx.Isolate().Dispose()
 	defer ctx.Close()
-	val, _ := ctx.RunScript("const foo = { bar: 'baz', '2': 2}; foo", "")
+	symIter := v8.SymbolIterator(ctx.Isolate())
+	val, _ := ctx.RunScript("const foo = { bar: 'baz', '2': 2, [Symbol.iterator]: 3}; foo", "")
 	obj, _ := val.AsObject()
 	if !obj.Has("bar") {
 		t.Error("expected property to exist")
@@ -186,9 +213,16 @@ func TestObjectDelete(t *testing.T) {
 	if obj.Has("bar") {
 		t.Error("expected property to be deleted")
 	}
+	if !obj.DeleteSymbol(symIter) {
+		t.Error("expected delete to return true, got false")
+	}
+	if obj.HasSymbol(symIter) {
+		t.Error("expected property to be deleted")
+	}
 	if !obj.DeleteIdx(2) {
 		t.Error("expected delete to return true, got false")
 	}
+
 }
 
 func ExampleObject_global() {
